+++ mordant-jvm-ffm/src/jvmMain/kotlin/com/github/ajalt/mordant/terminal/terminalinterface/ffm/TerminalInterfaceProvider.ffm.kt
@@ -7,13 +7,11 @@ import com.oracle.svm.core.annotate.TargetClass
 
 class TerminalInterfaceProviderFfm : TerminalInterfaceProvider {
     override fun load(): TerminalInterface? {
-        try {
-            if (!TerminalInterfaceProviderFfm::class.java.module.isNativeAccessEnabled()) {
-                return null
-            }
-        } catch (e: NoSuchMethodError) {
-           // isNativeAccessEnabled doesn't exist on old JDKs
-           return null
+        // The preview FFM API on JDKs older than 22 is incompatible
+        if (Runtime.version().feature() < 22
+            || !TerminalInterfaceProviderFfm::class.java.module.isNativeAccessEnabled()
+        ) {
+            return null
         }
         val os = System.getProperty("os.name")
         return when {