+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fixed
 - Fixed table border being drawn incorrectly in some tables that use rowspan. [(#272)](https://github.com/ajalt/mordant/issues/272)
 - Fixed GraalVM reflection configs listing the wrong class names for the terminal interface providers. [(#292)](https://github.com/ajalt/mordant/issues/292)
+- Fixed `ClassCastException` when creating a `Terminal` with the `mordant-jvm-ffm` module on Java 20 and 21. [(#278)](https://github.com/ajalt/mordant/issues/278)
 
 ## 3.0.2
 ### Changed